Make Keyman non blocking for slotted elements so they aren't hidden until Keyman inits
