+++ package.json
@@ -2,12 +2,11 @@
 	"name": "living-dictionaries",
 	"version": "0.0.1",
 	"scripts": {
-		"dev": "pnpm site",
-		"dev:open": "pnpm site --open",
-		"build": "pnpm site:build",
+		"dev": "pnpm --filter=site dev",
+		"dev:open": "pnpm --filter=site dev --open",
+		"build": "pnpm --filter=site build",
+		"preview": "pnpm --filter=site preview",
 		"test": "pnpm --filter=site... test",
-		"site": "pnpm --filter=site dev",
-		"site:build": "pnpm --filter=site build",
 		"site:test": "pnpm --filter=site test",
 		"parts": "pnpm --filter=parts dev",
 		"parts:build": "pnpm --filter=parts build",
