+++ packages/site/src/lib/components/entry/EditField.svelte
@@ -3,7 +3,6 @@
   import { createEventDispatcher } from 'svelte';
   import Button from 'svelte-pieces/ui/Button.svelte';
   import Keyman from '$lib/components/keyboards/keyman/Keyman.svelte';
-  import InputWrapper from '$lib/components/keyboards/keyman/InputWrapper.svelte';
   const dispatch = createEventDispatcher<{
     close: boolean;
     valueupdate: { field: string; newValue: string };
@@ -123,50 +122,38 @@
   <div class="rounded-md shadow-sm">
     {#if field === 'nt'}
       {#await import('@living-dictionaries/parts/src/lib/editor/ClassicCustomized.svelte') then { default: ClassicCustomized }}
-        <Keyman>
-          <InputWrapper
-            fixed
-            target=".ck-editor__editable_inline"
-            canChooseKeyboard
-            position="bottom">
-            <ClassicCustomized {editorConfig} bind:html={value} />
-          </InputWrapper>
+        <Keyman fixed target=".ck-editor__editable_inline" canChooseKeyboard position="bottom">
+          <ClassicCustomized {editorConfig} bind:html={value} />
         </Keyman>
       {/await}
     {:else if field.startsWith('gl') || field.startsWith('xs')}
-      <Keyman>
-        <InputWrapper fixed bcp={field.split('.')[1]}>
-          <input
-            bind:this={inputEl}
-            dir="ltr"
-            type="text"
-            required={field === 'lx'}
-            use:autofocus
-            bind:value
-            class:sompeng={display === 'Sompeng-Mardir'}
-            class="form-input block w-full pr-9" />
-        </InputWrapper>
+      <Keyman fixed bcp={field.split('.')[1]}>
+        <input
+          bind:this={inputEl}
+          dir="ltr"
+          type="text"
+          use:autofocus
+          bind:value
+          class:sompeng={display === 'Sompeng-Mardir'}
+          class="form-input block w-full pr-9" />
       </Keyman>
     {:else if field.startsWith('lo') || field === 'lx'}
-      <Keyman>
-        <InputWrapper fixed canChooseKeyboard>
-          <input
-            bind:this={inputEl}
-            dir="ltr"
-            type="text"
-            required={field === 'lx'}
-            use:autofocus
-            bind:value
-            class:sompeng={display === 'Sompeng-Mardir'}
-            class="form-input block w-full pr-9" />
-        </InputWrapper>
+      <Keyman fixed canChooseKeyboard>
+        <input
+          bind:this={inputEl}
+          dir="ltr"
+          type="text"
+          required={field === 'lx'}
+          use:autofocus
+          bind:value
+          class:sompeng={display === 'Sompeng-Mardir'}
+          class="form-input block w-full pr-9" />
       </Keyman>
     {:else}
       <input
         bind:this={inputEl}
         dir="ltr"
         type="text"
-        required={field === 'lx'}
         use:autofocus
         bind:value
         class="form-input block w-full" />
